refactor(core): split updater.rs (2.2k lines) into updater/{embedded,manifest,types,version,install,verify,services,rollback,http,tests} (#439)
Pure move; mod.rs re-exports the public surface so aifw_core::updater::*
paths used by api/cli/daemon are unchanged. Sibling items are pub(super)
and shared through private glob imports in mod.rs; include_str! paths
adjusted. The manifest .expect is now backed by embedded_manifest_parses,
so a malformed freebsd/manifest.json fails the test run, not the first
updater call on an appliance. Guard-test file lists updated.
